주문_기능_구현 : feat : 주문 상태 변경시 동시성 문제 해결, 카트 비어있을때 예외처리 https://github.com/CampusTable/campus-table-be-clone/issues/32

+++ src/main/java/shop/campustable/campustablebeclone/domain/order/repository/OrderItemRepository.java
@@ -3,6 +3,8 @@ package shop.campustable.campustablebeclone.domain.order.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import shop.campustable.campustablebeclone.domain.order.entity.OrderItem;
 
-public interface OrderItemRepository extends JpaRepository<OrderItem, Integer> {
+public interface OrderItemRepository extends JpaRepository<OrderItem, Long> {
+
+
 
 }
